[flang] Fix character length handling in array constructors
Array constructor item handling needed to be able to cope with
the fact that DynamicType has two ways of representing the length
of CHARACTER data (viz., a known constant, or a DeclTypeSpec's
ParamValue) and not require the items of an untyped array constructor
to have DynamicType values that compare equal via its operator==().

Fixes f18-stage bug !105.