[flang] Fold more reduction intrinsic function calls
Refactor the recently-implemented MAXVAL/MINVAL so that
the parts that can be used to implement other reduction
transformational intrinsic function folding are exposed.

Use them to implement folding of IALL, IANY, IPARITY,
SUM. and PRODUCT.  Replace the folding of ALL & ANY to
use the new infrastructure and become able to handle DIM=
arguments.