[flang] Include default component initialization in static initializers
The combined initializers constructed from DATA statements and explicit
static initialization in declarations needs to include derived type
component default initializations, overriding those default values
without complaint with values from explicit DATA statement or declaration
initializations when they overlap.  This also has to work for objects
with storage association due to EQUIVALENCE.  When storage association causes
default component initializations to overlap, emit errors if and only
if the values differ (See Fortran 2018 subclause 19.5.3, esp. paragraph
10).

The f18 front-end has a module that analyzes and converts DATA statements
into equivalent static initializers for objects.  For storage-associated
objects, compiler-generated objects are created that overlay the entire
association and fill it with a combined initializer.  This "data-to-inits"
module already exists, and this patch is essentially extension and
clean-up of its machinery to complete the job.

This patch also includes code to create explicit initializers for static
objects that don't appear in DATA or have explicit static initialization
and are not storage-associated with such things, but whose derived types
have default component initialization.  We might remove that code if it
turns out to be better to figure this problem out in lowering, but it was
trivial to do here with the new machinery for solving the more general
problem.

Also: emit EQUIVALENCE to module files; mark compiler-created symbols
and *don't* emit those to module files.