[flang] Signal EOR in non advancing IO and move to next record
When an end of record is met in non advancing IO:
- Set IOSTAT if present according to 12.11.4 (5).
- Position the file to the next record (12.11.4 (4)).

The previous code was only signaling EOR for fixed record length IO.
Reading at 12.11.4, I do not find the rational for this condition, so I
removed it.
It also does not seems the presence of padding should also not prevent
the EOR signaling.

EOR in input is not an error, I am not confident to modify
ErrorHandler.isError() here, but in FinishReadingRecord, the code should
not bailed if the error is simply an end of record.

I did not check the SIZE requirements here because GetSize runtime is
not yet implemented.