[flang] Make f18 build with GCC 11.2.0 and with --std=c++20
C++20 deprecates the implicit access to the value of "this" in
a lambda closure with "[=]"; rewrite some as "[this]" or "[=, this]".

Address a legitimate, but benign, warning about overindexing when
converting REAL(10) 80-bit x87 floating-point constants to host
long doubles (when available), which GCC pads out to 16 bytes.

Add an explicit no-op constructor to Fortran::parser::Verbatim.
(It has worked fine without one until now, but now it seems necessary.)

Builds were verified out-of-tree against LLVM built with GCC 9.3.0
in shared library mode.