[flang] add an evaluate::IsAllocatable() helper
Previously, some tests for the allocatable attributes where relying on
the last symbol of a designator  being an allocatable.

This is wrong since this was considering substrings and array section of
allocatables as being allocatable, which is not the case (see NOTE 2 in
Fortran 2018 section 9.5.3).

Add some semantic tests for ALLOCATED to test the newly added helper.
Note that ifort and nag are rejecting coindexed-named-object in
ALLOCATED (`allocated(coarray_scalar_alloc[2])`).
I think it is wrong given allocated argument is intent(in) as per
16.2.1 point 3.
So 15.5.2.6 point 4 regarding allocatable dummy is not violated (If the actual
argument is a coindexed object, the dummy argument shall have the INTENT (IN)
attribute.) and I think this is valid. gfortran accepts it.