[flang] Local generics must not shadow host-associated generics
It is possible for generic interfaces of equivalent (but not necessarily
identical -- operator(.eq.) is equivalent to operator(==)) names to
be declared in a host scope and a nested scope, and the nested declaration
should function as an extension of the host's.

This MR comprises two commits -- the first is the patch that fixes bug #457 and
the second builds upon it.  I have not yet decided whether to squash them into
one patch or to apply them in sequence.  You only need to consider the latest
commit, which shows up *first* in the list below.