[flang] Undo lowering part of GetExpr() change (f18-stage ONLY, not for upstreaming)
A change to GetExpr() here in f18-stage fixed a bug in semantics
and alo included an API change that rippled through lowering.
I recently integrated the fix into llvm-project without the
API change that affected lowering -- it was discretionary and
not necessary for the fix -- and now llvm-project is out of
synch with f18-stage.  Revert the lowering part of the change
here to get things back into concordance between the two trees.