Fix a long-standing bug wrt BIND(C) and procedures.
Substitute the BIND(C) name early and skip any name uniquing on the
assumption that BIND(C) names will actually be unique. This is
sufficient for legal programs and does not use unique names uselessly
backwards as before.

Preserves the functionality of flagging functions with a BIND(C) name
and capturing the name in a like-minded attribute. This is done should
identifying subprograms with BIND(C) be needed and/or if uniquing and
rewriting these names should become desired.

Add a test for Fortran namesapce collisions and BIND(C) overrides.