Emit a fatal error when the front end passes subprograms with identical
* Emit a fatal error when the front end passes subprograms with identical
names and different definitions.

Flang catches some, not all, cases where multiple subprograms have the
same Fortran name but differing BIND(C) names. NAG compiles such
programs without error.

  subroutine s ...
  subroutine s() BIND(C, Name='not_s') ...

In cases where flang does not flag this situation as an error in
semantics, it also drops the BIND(C) name on the floor and it is no
present in the Se::Symbol.