Fix assumed size in alternate entry points
Assumed sized were wrongly categorized as "constant shape" in BoxAnalyzer.h,
this had no impact so far because they are dummies, so no allocation is
attempted for them when lowering the symbol... except for alternate
entries where a mock allocation was attempted for them in the entries
where the symbol is not in the argument list.

It seems wrong to consider assumed size arrays as having constant shape,
consider them as "dynamic" (this has a small ripple effects in some
tests because the constant extents are now "validated" to be greater
or equal to zero, llvm just folds this).