[flang][runtime] Improve G0 output editing
G0 output editing should never overflow an output field and fill it
with asterisks.  It should also never elide the "E" in an exponent
field, even if it has more than three digits.

This is for nag_hibiya tests c0l and c0n.  The output still does not match NAG's but
I think I'm happy with it.